pyproject.toml
[tool.poetry]
name = "r11data"
version = "0.1.0"
description = ""
authors = ["Lukas Plank <lupl@tuta.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
rdflib = "^7.0.0"
rich = "^13.7.1"
tabulardf = "^0.1.1"
convertdate = "^2.4.0"
pydantic = "^2.8.2"
toolz = "^0.12.1"
sparqlwrapper = "^2.0.0"
lodkit = "^0.1.13"
pytz = "^2024.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
